+++ src/main/java/com/guiferrini/CasaCodigo/model/CuponDTO.java
@@ -1,5 +1,6 @@
 package com.guiferrini.CasaCodigo.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import javax.validation.constraints.Future;
@@ -17,7 +18,7 @@ public class CuponDTO {
 	
 	@NotNull(message = "A porcentagem do Desconto é obrigatoria e positiva.")
 	@Positive
-	private Double desconto;
+	private BigDecimal desconto;
 	
 	@NotNull(message = "A Data é Obrigatórioa e no Futuro.")
 	@Future
@@ -29,7 +30,7 @@ public class CuponDTO {
 	}
 
 	public CuponDTO(@NotBlank(message = "O Código é Obrigatorio.") String codigo, 
-			@NotNull(message = "A porcentagem do Desconto é obrigatoria e positiva.") @Positive Double desconto, 
+			@NotNull(message = "A porcentagem do Desconto é obrigatoria e positiva.") @Positive BigDecimal desconto, 
 			@NotNull(message = "A Data é Obrigatórioa e no Futuro.") @Future LocalDate validade) {
 		super();
 		this.codigo = codigo;
@@ -41,7 +42,7 @@ public class CuponDTO {
 		return codigo;
 	}
 
-	public Double getDesconto() {
+	public BigDecimal getDesconto() {
 		return desconto;
 	}
