+++ src/main/java/com/guiferrini/CasaCodigo/model/NovoLivroDTO.java
@@ -1,5 +1,6 @@
 package com.guiferrini.CasaCodigo.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import javax.persistence.EntityManager;
@@ -28,7 +29,7 @@ public class NovoLivroDTO {
 
 	@NotNull(message="Preço obrigatório")
 	@Min(20)
-	private Double preco;
+	private BigDecimal preco;
 	
 	@NotNull(message="Número de paginas obrigatório")
 	@Min(100)
@@ -58,7 +59,7 @@ public class NovoLivroDTO {
  
 	public NovoLivroDTO(@NotBlank String titulo,
 			@NotBlank @Size(max = 500) String resumo, String sumario,
-			@NotNull @Min(20) Double preco,
+			@NotNull @Min(20) BigDecimal preco,
 			@NotNull @Min(100) Integer paginas,
 			@NotBlank String identificador, @Future
 			@NotNull LocalDate date, @NotNull String idCategoria,
@@ -99,11 +100,11 @@ public class NovoLivroDTO {
 		this.sumario = sumario;
 	}
 
-	public Double getPreco() {
+	public BigDecimal getPreco() {
 		return preco;
 	}
 
-	public void setPreco(Double preco) {
+	public void setPreco(BigDecimal preco) {
 		this.preco = preco;
 	}
